Move `Error` into its own module (internal organization only - public API remains).
