+++ src/backup.rs
@@ -36,7 +36,8 @@ use std::time::Duration;
 
 use ffi;
 
-use {DatabaseName, Connection, Error, Result};
+use {DatabaseName, Connection, Result};
+use error::{error_from_sqlite_code, error_from_handle};
 
 impl Connection {
     /// Back up the `name` database to the given destination path.
@@ -70,8 +71,8 @@ impl Connection {
 
         match r {
             Done => Ok(()),
-            Busy => Err(Error::from_handle(ptr::null_mut(), ffi::SQLITE_BUSY)),
-            Locked => Err(Error::from_handle(ptr::null_mut(), ffi::SQLITE_LOCKED)),
+            Busy => Err(error_from_handle(ptr::null_mut(), ffi::SQLITE_BUSY)),
+            Locked => Err(error_from_handle(ptr::null_mut(), ffi::SQLITE_LOCKED)),
             More => unreachable!(),
         }
     }
@@ -115,8 +116,8 @@ impl Connection {
 
         match r {
             Done => Ok(()),
-            Busy => Err(Error::from_handle(ptr::null_mut(), ffi::SQLITE_BUSY)),
-            Locked => Err(Error::from_handle(ptr::null_mut(), ffi::SQLITE_LOCKED)),
+            Busy => Err(error_from_handle(ptr::null_mut(), ffi::SQLITE_BUSY)),
+            Locked => Err(error_from_handle(ptr::null_mut(), ffi::SQLITE_LOCKED)),
             More => unreachable!(),
         }
     }
@@ -201,7 +202,7 @@ impl<'a, 'b> Backup<'a, 'b> {
                                              from.db.borrow_mut().db,
                                              from_name.as_ptr());
             if b.is_null() {
-                return Err(Error::from_handle(to_db, ffi::sqlite3_errcode(to_db)));
+                return Err(error_from_handle(to_db, ffi::sqlite3_errcode(to_db)));
             }
             b
         };
@@ -244,7 +245,7 @@ impl<'a, 'b> Backup<'a, 'b> {
             ffi::SQLITE_OK => Ok(More),
             ffi::SQLITE_BUSY => Ok(Busy),
             ffi::SQLITE_LOCKED => Ok(Locked),
-            _ => Err(Error::from_sqlite_code(rc, None)),
+            _ => Err(error_from_sqlite_code(rc, None)),
         }
     }
