+++ src/trace.rs
@@ -8,7 +8,8 @@ use std::str;
 use std::time::Duration;
 
 use super::ffi;
-use {Error, Result, Connection};
+use {Result, Connection};
+use error::error_from_sqlite_code;
 
 /// Set up the process-wide SQLite error logging callback.
 /// This function is marked unsafe for two reasons:
@@ -45,7 +46,7 @@ pub unsafe fn config_log(callback: Option<fn(c_int, &str)>) -> Result<()> {
     if rc == ffi::SQLITE_OK {
         Ok(())
     } else {
-        Err(Error::from_sqlite_code(rc, None))
+        Err(error_from_sqlite_code(rc, None))
     }
 }
